Keep extra metadata in submit (#12)
To be able to keep data such as links, creation-dates etc. from the
content submit`ed, we need to make a copy of the post data, remove the
actual content and post this to the work queue of Scio.

The main usage area for this is the "creation-date" and "link"/"links"
fields kept from the feed download scripts.